# main/resources/swagger/avia-client.yaml
openapi: 3.1.0
servers:
  - url: http://localhost:8080/api/v1/avia
    description: Local server
info:
  title: Flight Search API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
...
  /createBooking:
    post:
      summary: Создание бронирования
      operationId: createBooking
      tags:
        - Avia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequestDto'
      responses:
        '200':
          description: Успешное создание бронирования
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBookingResponseDto'
  /cancelBooking:
    post:
      summary: Отмена бронирования
      operationId: cancelBooking
      tags:
        - Avia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequestDto'
      responses:
        '200':
          description: Успешная отмена бронирования
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelBookingResponseDto'
...
components:
  schemas:
...
    PassengerType:
      type: string
      enum: [ADULT, CHILD, INFANT]
...
    SimpleTextReferenceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    FlightTokenDto:
      type: object
      properties:
        service:
          type: string
        connector:
          type: string
        searchInfo:
          type: string
    MoneyDto:
      type: object
      properties:
        amount:
          type: number
        currency:
          $ref: '#/components/schemas/CurrencyDto'
      required:
        - amount
        - currency
    CurrencyDto:
      type: string
      enum:
        - AED  # UAE Dirham
        - AFN  # Afghani
        - ALL  # Lek
        - AMD  # Armenian Dram
        - ANG  # Netherlands Antillean Guilder
        - AOA  # Kwanza
        - ARS  # Argentine Peso
        - AUD  # Australian Dollar
        - AWG  # Aruban Florin
        - AZN  # Azerbaijan Manat
        - BAM  # Convertible Mark
        - BBD  # Barbados Dollar
        - BDT  # Taka
        - BGN  # Bulgarian Lev
        - BHD  # Bahraini Dinar
        - BIF  # Burundi Franc
        - BMD  # Bermudian Dollar
        - BND  # Brunei Dollar
        - BOB  # Boliviano
        - BOV  # Mvdol
        - BRL  # Brazilian Real
        - BSD  # Bahamian Dollar
        - BTN  # Ngultrum
        - BWP  # Pula
        - BYN  # Belarusian Ruble
        - BZD  # Belize Dollar
        - CAD  # Canadian Dollar
        - CDF  # Congolese Franc
        - CHE  # WIR Euro
        - CHF  # Swiss Franc
        - CHW  # WIR Franc
        - CLF  # Unidad de Fomento
        - CLP  # Chilean Peso
        - CNY  # Yuan Renminbi
        - COP  # Colombian Peso
        - COU  # Unidad de Valor Real
        - CRC  # Costa Rican Colon
        - CUC  # Peso Convertible
        - CUP  # Cuban Peso
        - CVE  # Cabo Verde Escudo
        - CZK  # Czech Koruna
        - DJF  # Djibouti Franc
        - DKK  # Danish Krone
        - DOP  # Dominican Peso
        - DZD  # Algerian Dinar
        - EGP  # Egyptian Pound
        - ERN  # Nakfa
        - ETB  # Ethiopian Birr
        - EUR  # Euro
        - FJD  # Fiji Dollar
        - FKP  # Falkland Islands Pound
        - GBP  # Pound Sterling
        - GEL  # Lari
        - GHS  # Ghana Cedi
        - GIP  # Gibraltar Pound
        - GMD  # Dalasi
        - GNF  # Guinean Franc
        - GTQ  # Quetzal
        - GYD  # Guyana Dollar
        - HKD  # Hong Kong Dollar
        - HNL  # Lempira
        - HTG  # Gourde
        - HUF  # Forint
        - IDR  # Rupiah
        - ILS  # New Israeli Sheqel
        - INR  # Indian Rupee
        - IQD  # Iraqi Dinar
        - IRR  # Iranian Rial
        - ISK  # Iceland Krona
        - JMD  # Jamaican Dollar
        - JOD  # Jordanian Dinar
        - JPY  # Yen
        - KES  # Kenyan Shilling
        - KGS  # Som
        - KHR  # Riel
        - KMF  # Comorian Franc
        - KPW  # North Korean Won
        - KRW  # Won
        - KWD  # Kuwaiti Dinar
        - KYD  # Cayman Islands Dollar
        - KZT  # Tenge
        - LAK  # Lao Kip
        - LBP  # Lebanese Pound
        - LKR  # Sri Lanka Rupee
        - LRD  # Liberian Dollar
        - LSL  # Loti
        - LYD  # Libyan Dinar
        - MAD  # Moroccan Dirham
        - MDL  # Moldovan Leu
        - MGA  # Malagasy Ariary
        - MKD  # Denar
        - MMK  # Kyat
        - MNT  # Tugrik
        - MOP  # Pataca
        - MRU  # Ouguiya
        - MUR  # Mauritius Rupee
        - MVR  # Rufiyaa
        - MWK  # Malawi Kwacha
        - MXN  # Mexican Peso
        - MXV  # Mexican Unidad de Inversion (UDI)
        - MYR  # Malaysian Ringgit
        - MZN  # Mozambique Metical
        - NAD  # Namibia Dollar
        - NGN  # Naira
        - NIO  # Cordoba Oro
        - NOK  # Norwegian Krone
        - NPR  # Nepalese Rupee
        - NZD  # New Zealand Dollar
        - OMR  # Rial Omani
        - PAB  # Balboa
        - PEN  # Sol
        - PGK  # Kina
        - PHP  # Philippine Peso
        - PKR  # Pakistan Rupee
        - PLN  # Zloty
        - PYG  # Guarani
        - QAR  # Qatari Rial
        - RON  # Romanian Leu
        - RSD  # Serbian Dinar
        - RUB  # Russian Ruble
        - RWF  # Rwanda Franc
        - SAR  # Saudi Riyal
        - SBD  # Solomon Islands Dollar
        - SCR  # Seychelles Rupee
        - SDG  # Sudanese Pound
        - SEK  # Swedish Krona
        - SGD  # Singapore Dollar
        - SHP  # Saint Helena Pound
        - SLE  # Leone
        - SLL  # Leone
        - SOS  # Somali Shilling
        - SRD  # Surinam Dollar
        - SSP  # South Sudanese Pound
        - STN  # Dobra
        - SVC  # El Salvador Colon
        - SYP  # Syrian Pound
        - SZL  # Lilangeni
        - THB  # Baht
        - TJS  # Somoni
        - TMT  # Turkmenistan New Manat
        - TND  # Tunisian Dinar
        - TOP  # Pa’anga
        - TRY  # Turkish Lira
        - TTD  # Trinidad and Tobago Dollar
        - TWD  # New Taiwan Dollar
        - TZS  # Tanzanian Shilling
        - UAH  # Hryvnia
        - UGX  # Uganda Shilling
        - USD  # US Dollar
        - USN  # US Dollar (Next day)
        - UYI  # Uruguay Peso en Unidades Indexadas (UI)
        - UYU  # Peso Uruguayo
        - UYW  # Unidad Previsional
        - UZS  # Uzbekistan Sum
        - VED  # Bolívar Soberano
        - VES  # Bolívar Soberano
        - VND  # Dong
        - VUV  # Vatu
        - WST  # Tala
        - XAF  # CFA Franc BEAC
        - XAG  # Silver
        - XAU  # Gold
        - XBA  # EURCO
        - XBB  # EMU-6
        - XBC  # EUA-9
        - XBD  # EUA-17
        - XCD  # East Caribbean Dollar
        - XDR  # Special Drawing Right
        - XOF  # CFA Franc BCEAO
        - XPD  # Palladium
        - XPF  # CFP Franc
        - XPT  # Platinum
        - XSU  # Sucre
        - XTS  # Testing code
        - XUA  # ADB Unit of Account
        - XXX  # No currency
        - YER  # Yemeni Rial
        - ZAR  # Rand
        - ZMW  # Zambian Kwacha
        - ZWL  # Zimbabwe Dollar
        - HRK  # Croatian Kuna
        - ZWD  # Zimbabwean dollar
        - BYR  # Belarusian Ruble (old)
    FlightLegDto:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        duration:
          $ref: '#/components/schemas/DurationDto'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/FlightSegmentDto'
      required:
        - duration
        - segments
        - uid
    DurationDto:
      type: object
      properties:
        days:
          type: integer
        hours:
          type: integer
        minutes:
          type: integer
    FlightSegmentDto:
      type: object
      properties:
        departureAirport:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        departureCity:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        departureCountry:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        departureDateTime:
          type: string
          format: date-time
        departureTerminal:
          type: string

        arrivalAirport:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        arrivalCity:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        arrivalCountry:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        arrivalDateTime:
          type: string
          format: date-time
        arrivalTerminal:
          type: string

        flightNumber:
          type: string
        carrierTripNumber:
          type: string

        serviceClass:
          $ref: '#/components/schemas/FlightClassDto'

        flightDuration:
          $ref: '#/components/schemas/DurationDto'
        transferDuration:
          $ref: '#/components/schemas/DurationDto'

        validatingCarrier:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        operatingCarrier:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        aircraft:
          $ref: '#/components/schemas/SimpleTextReferenceDto'

        seatsAvailable:
          type: integer
        baggageAvailable:
          type: boolean
        baggageInfo:
          $ref: '#/components/schemas/BaggageInfoDto'
        cabinBaggageInfo:
          $ref: '#/components/schemas/BaggageInfoDto'

        techStops:
          type: array
          items:
            $ref: '#/components/schemas/TechStopDto'
      required:
        - departureAirport
        - departureCity
        - departureCountry
        - departureDateTime
        - arrivalAirport
        - arrivalCity
        - arrivalCountry
        - arrivalDateTime
        - flightNumber
        - serviceClass
        - flightDuration
        - transferDuration
        - validatingCarrier
        - operatingCarrier
        - aircraft
        - seatsAvailable
        - baggageAvailable
        - baggageInfo
        - cabinBaggageInfo
        - techStops
    FlightClassDto:
      type: string
      enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, PREMIUM_BUSINESS, FIRST, PREMIUM_FIRST, SUBSIDIZED]
    BaggageInfoDto:
      type: object
      properties:
        piecesAvailable:
          type: integer
        weightPerPiece:
          type: integer
        dimensions:
          $ref: '#/components/schemas/BaggageDimensionsDto'
    BaggageDimensionsDto:
      type: object
      properties:
        width:
          type: integer
        length:
          type: integer
        height:
          type: integer
    TechStopDto:
      type: object
      properties:
        airport:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        city:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        country:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        arrivalDateTime:
          type: string
          format: date-time
        departureDateTime:
          type: string
          format: date-time
        duration:
          $ref: '#/components/schemas/DurationDto'
      required:
        - airport
        - city
        - country
        - arrivalDateTime
        - departureDateTime
        - duration
...
    CreateBookingRequestDto:
      type: object
      required:
        - selectFlightToken
        - payer
        - passengers
        - orderUid
        - confirmedPrice
      properties:
        selectFlightToken:
          $ref: '#/components/schemas/FlightTokenDto'
        payer:
          $ref: '#/components/schemas/BookingPayerDto'
        passengers:
          type: array
          items:
            $ref: '#/components/schemas/BookingPassengerDto'
        orderUid:
          type: string
        confirmedPrice:
          type: string
          format: number
    BookingPayerDto:
      type: object
      required:
        - email
        - phone
        - fullName
      properties:
        email:
          type: string
        phone:
          type: string
        fullName:
          type: string
    BookingPassengerDto:
      type: object
      required:
        - firstName
        - lastName
        - birthDate
        - gender
        - document
        - citizenship
        - type
        - personHasNoMiddleName
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 40
          pattern: "^(?![.-])(?!.*[.-]{2})(?!.*[.-]$)[A-Za-zА-Яа-яёЁ.\\-]+$"
          description: Имя пассажира
        middleName:
          type: string
          minLength: 2
          maxLength: 40
          pattern: "^(?![.-])(?!.*[.-]{2})(?!.*[.-]$)[A-Za-zА-Яа-яёЁ.\\-]+$"
          description: Отчество пассажира (обязательное, если personHasNoMiddleName = false)
        personHasNoMiddleName:
          type: boolean
          default: false
          description: true — если отчество отсутствует
        lastName:
          type: string
          minLength: 2
          maxLength: 40
          pattern: "^(?![.-])(?!.*[.-]{2})(?!.*[.-]$)[A-Za-zА-Яа-яёЁ.\\-]+$"
          description: Фамилия пассажира
        birthDate:
          type: string
          format: date
          description: Дата рождения пассажира (YYYY-MM-DD)
        gender:
          type: string
          enum: [ MALE, FEMALE ]
          description: Пол пассажира
        document:
          $ref: '#/components/schemas/PassengerDocumentInfoDto'
        citizenship:
          $ref: '#/components/schemas/CitizenshipDto'
        type:
          $ref: '#/components/schemas/PassengerType'
        aviaBonusCards:
          type: array
          items:
            $ref: '#/components/schemas/AviaBonusCardDto'
        additionalBaggagePieces:
          type: integer
          minimum: 0
    PassengerDocumentInfoDto:
      type: object
      required:
        - type
        - number
        - issuedDate
      properties:
        type:
          type: string
          enum: [ RUSSIAN_PASSPORT, RUSSIAN_FOREIGN_PASSPORT ]
        number:
          type: string
          description: |
            Формат зависит от типа:
            - RUSSIAN_PASSPORT: XXXX XXXXXX
            - RUSSIAN_FOREIGN_PASSPORT: XX XXXXXXX
        issuedDate:
          type: string
          format: date
          description: Дата выдачи документа (YYYY-MM-DD)
        expiryDate:
          type: string
          format: date
          description: |
            Дата окончания действия документа.
            Обязательна, если type = RUSSIAN_FOREIGN_PASSPORT
    CitizenshipDto:
      type: object
      required:
        - uid
        - name
      properties:
        uid:
          type: string
          format: uuid
        name:
          type: string
    AviaBonusCardDto:
      type: object
      required:
        - number
        - iataCode
      properties:
        number:
          type: string
          description: Номер бонусной карты
        iataCode:
          type: string
          pattern: '^[A-Z]{2}$'
          description: IATA-код авиакомпании (2 буквы)
    CreateBookingResponseDto:
      type: object
      required:
        - orderUid
        - currentTime
        - aviaProduct
      properties:
        orderUid:
          type: string
        currentTime:
          type: string
          format: date-time
        aviaProduct:
          $ref: '#/components/schemas/AviaProductDto'
    AviaProductDto:
      type: object
      required:
        - type
        - externalOrderId
        - externalBillingNumber
        - limitTimeForPay
        - flight
        - aviaProductPrice
        - provider
        - isComplexRoute
        - bookingTimeLimitExpired
        - nextStatusInProcess
        - tickets
        - exchanged
        - uid
        - name
        - status
        - allowModification
      properties:
        type:
          type: string
          default: AviaProductDto
        externalOrderId:
          type: string
        externalBillingNumber:
          type: string
        limitTimeForPay:
          type: string
          format: date-time
        flight:
          $ref: '#/components/schemas/AviaFlightDto'
        aviaProductPrice:
          $ref: '#/components/schemas/AviaProductPriceDto'
        provider:
          $ref: '#/components/schemas/ProviderAndService'
        isComplexRoute:
          type: boolean
        bookingTimeLimitExpired:
          type: boolean
        nextStatusInProcess:
          $ref: '#/components/schemas/ProductStatus'
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/TicketDto'
        customAttributes:
          type: object
          property:
            $ref: '#/components/schemas/AviaProductAttributeType'
          additionalProperties:
            type: string
        exchanged:
          type: boolean
        uid:
          type: string
          format: uuid
          description: Уникальный идентификатор продукта
        name:
          type: string
          description: Название продукта
        status:
          $ref: '#/components/schemas/ProductStatus'
        allowModification:
          type: boolean
          description: Признак разрешения на модификацию
        approvalEventStatus:
          $ref: '#/components/schemas/PersonEventStatus'
        approvalEventUid:
          type: string
          format: uuid
        isSystemCreated:
          type: boolean
          nullable: true
          description: Признак системного создания
        providerForAccounting:
          $ref: '#/components/schemas/ProviderAndService'
        personUids:
          type: array
          uniqueItems: true
          items:
            type: string
        cities:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/CityDto'
    CityDto:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        name:
          type: string
    AviaProductAttributeType:
      type: string
      enum:
        - VIP_SERVICE_RESERVATION_TOKEN
      description: Тип кастомного атрибута авиапродукта
    AviaFlightDto:
      type: object
      required:
        - duration
        - validatingCarrier
        - additionalFaresAvailable
        - refundAvailable
        - exchangeAvailable
        - baggageAvailable
        - fareFamilyType
        - fareFamilyMarketingName
        - legs
        - baggageInfo
        - cabinBaggageInfo
      properties:
        duration:
          $ref: '#/components/schemas/DurationDto'
        validatingCarrier:
          $ref: '#/components/schemas/SimpleTextReferenceDto'
        additionalFaresAvailable:
          type: boolean
        refundAvailable:
          type: boolean
        exchangeAvailable:
          type: boolean
        baggageAvailable:
          type: boolean
        fareFamilyType:
          type: string
        fareFamilyMarketingName:
          type: string
        legs:
          type: array
          items:
            $ref: '#/components/schemas/FlightLegDto'
        baggageInfo:
          $ref: '#/components/schemas/BaggageInfoDto'
        cabinBaggageInfo:
          $ref: '#/components/schemas/BaggageInfoDto'
    AviaProductPriceDto:
      type: object
      required:
        - providerPrice
        - serviceFee
        - totalPrice
        - refundAmount
        - totalPayedMinusTotalReturned
        - totalPayedMinusTotalReturnedMinusServiceFee
      properties:
        providerPrice:
          $ref: '#/components/schemas/MoneyDto'
        serviceFee:
          $ref: '#/components/schemas/MoneyDto'
        totalPrice:
          $ref: '#/components/schemas/MoneyDto'
        refundAmount:
          $ref: '#/components/schemas/MoneyDto'
        totalPayedMinusTotalReturned:
          $ref: '#/components/schemas/MoneyDto'
        totalPayedMinusTotalReturnedMinusServiceFee:
          $ref: '#/components/schemas/MoneyDto'
    ProviderAndService:
      type: string
      enum:
        - MYAGENT_AVIA
        - MYAGENT_HOTEL
        - MYAGENT_RZD
        - OSTROVOK_HOTEL
        - IM_RZD
        - VIPSERVICE_AVIA
        - IWAY_TRANSFER
        - DADATA_DICTIONARY
        - UNKNOWN
    PersonEventStatus:
      type: string
      description: Статус события согласования
      enum:
        - NEW         # На согласовании
        - APPROVED    # Согласовано
        - REJECTED    # Отклонено
        - CANCELLED   # Аннулировано
        - PAYMENT_EXPIRED  # Время на оплату истекло
        - PAID        # Оплачено
        - REAPPROVE   # На пересогласовании
        - APPROVAL_EXPIRED  # Время на согласование истекло
    ProductStatus:
      type: string
      enum:
        - NEW
        - BOOKING_CREATED
        - ISSUED
        - BOOKING_CANCELED
        - ERROR
        - REFUND
        - BOOKING_EXPIRED
        - EXCHANGE_REQUESTED
        - EXCHANGE
    TicketDto:
      type: object
      required:
        - uid
        - number
        - status
        - ticketPrice
        - passengers
      properties:
        uid:
          type: string
          format: uuid
        number:
          type: string
        status:
          $ref: '#/components/schemas/AviaTicketStatus'
        nextStatusInProcess:
          $ref: '#/components/schemas/AviaTicketStatus'
        ticketPrice:
          $ref: '#/components/schemas/AviaProductPriceDto'
        passengers:
          type: array
          items:
            $ref: '#/components/schemas/BookingPassengerDto'
        minimalPrice:
          $ref: '#/components/schemas/MoneyDto'
        pnr:
          type: string
        documentPath:
          type: string
        documentUid:
          type: string
        priceWasChanged:
          type: boolean
        externalId:
          $ref: '#/components/schemas/TicketExternalIdDto'
        ticketWay:
          $ref: '#/components/schemas/TicketWayDto'
        legUids:
          type: array
          uniqueItems: true
          items:
            type: string
    TicketWayDto:
      type: string
      enum:
        - WAY_THERE
        - WAY_BACK
        - ROUND_TRIP
        - COMPLEX_ROUTE
    TicketExternalIdDto:
      type: object
      properties:
        productUid:
          type: string
        locator:
          type: string
        passengerIds:
          type: array
          items:
            type: string
        passengerKeys:
          type: array
          items:
            type: string
        passengerUuids:
          type: array
          items:
            type: string
    AviaTicketStatus:
      type: string
      enum:
        - BOOKING_CREATED
        - BOOKING_CANCELED
        - BOOKING_EXPIRED
        - BOOKING_ISSUED
        - ISSUED
        - REFUND
        - ERROR
        - EXCHANGE_REQUESTED
        - EXCHANGE
    CancelBookingRequestDto:
      type: object
      required:
        - aviaProduct
      properties:
        aviaProduct:
          $ref: '#/components/schemas/AviaProductDto'
    CancelBookingResponseDto:
      type: object
      required:
        - aviaProduct
      properties:
        aviaProduct:
          $ref: '#/components/schemas/AviaProductDto'
...